+++ udata/tests/plugin.py
@@ -1,6 +1,5 @@
 import pytest
 import shlex
-import sys
 
 from contextlib import contextmanager
 from urllib.parse import urlparse
@@ -101,7 +100,16 @@ def get_settings(request):
     marker = request.node.get_closest_marker('settings')
     if marker:
         return marker.args[0]
-    return getattr(request.cls, 'settings', settings.Testing)
+    _settings = getattr(request.cls, 'settings', settings.Testing)
+    # apply the options(plugins) marker from pytest_flask as soon as app is created
+    # https://github.com/pytest-dev/pytest-flask/blob/a62ea18cb0fe89e3f3911192ab9ea4f9b12f8a16/pytest_flask/plugin.py#L126
+    # this lets us have default settings for plugins applied while testing
+    plugins = getattr(_settings, 'PLUGINS', [])
+    for options in request.node.iter_markers('options'):
+        option = options.kwargs.get('plugins', []) or options.kwargs.get('PLUGINS', [])
+        plugins += option
+    setattr(_settings, 'PLUGINS', plugins)
+    return _settings
 
 
 def drop_db(app):